connec to mongoDB
in config: i added the mongo atlas cluster link to our env folder, then i created a db file in the config folder, in this file i required mongoose, an abstraction layer to deal with the database when i create models, and created the connectDB function to export and connect to the server.js. made an object of info to pass in to avoid deprecation warnings.

in server.js: i required the db folder & called the connectDB() function. i also created a function to handle all unhandled promise rejections (or errors in connecting to the database). This function logs the error and closes the server when there is a mongo error. I don't want my app to run if there are issues with the server.
 master